CI: a C++ compiler on Rocky, and a float Clang would not let past
Windows built the C++ first time. The other two had one problem each, and
between them they make the case for building on more than one compiler.

Rocky had no C++ compiler at all - the dnf line installed gcc and the
project had not needed gcc-c++ until now.

macOS caught a real defect that GCC waved through: SDL_fabs takes a double
and the pick test was handing it floats, so they promoted silently.
-Wdouble-promotion is in the warning set precisely to catch that, and GCC
does not diagnose the argument-passing case while Clang does. With -Werror
that difference is a broken build on one platform and a clean one on
three, which is the exact shape of bug the strict-warnings discipline is
supposed to prevent - and it only surfaced because CI builds with both.

Fixed by widening explicitly rather than by reaching for SDL_fabsf: the
comparison wants double precision, and saying so is clearer than picking a
narrower function and hoping the reader knows why.

Ubuntu now also pins CMAKE_CXX_COMPILER to g++-14 rather than letting it
default to whichever g++ the runner image ships. It worked either way, but
a C compiler and a C++ compiler from different releases is a puzzle
waiting for somebody, and one line removes it.

Everything here was verified against a full local Clang build before
pushing, which is cheaper than finding it one CI round trip at a time.

+++ tests/test_render.c
@@ -342,7 +342,12 @@ TEST(picking_a_pixel_finds_the_ground_that_was_drawn_there) {
             bool on = gs_iso_pick(&cam, &t, sx, sy, &px, &py);
             CHECK(on);
 
-            double err = SDL_fabs(px - wx) + SDL_fabs(py - wy);
+            // Widened explicitly. SDL_fabs takes a double, so passing floats
+            // promotes them silently - which GCC lets past and Clang does not,
+            // and -Werror means that difference is a broken build on one
+            // platform only.
+            double err = SDL_fabs((double)px - (double)wx) +
+                         SDL_fabs((double)py - (double)wy);
             if (err > worst) worst = err;
         }
     }
